+++ exp/README.md
@@ -1,0 +1,4 @@
+# Experimental Packages
+
+These packages do not have the same stability guarantee as the rest of the tools
+in this repository. If you use these, vendor them aggressively.
