+++ exp/gen/typemap/testdata/root_pkg.proto
