+++ protoc-gen-twirp/generator.go
@@ -29,9 +29,9 @@ import (
 	"github.com/golang/protobuf/protoc-gen-go/descriptor"
 	plugin "github.com/golang/protobuf/protoc-gen-go/plugin"
 	"github.com/pkg/errors"
-	"github.com/twitchtv/twirp/internal/gen"
-	"github.com/twitchtv/twirp/internal/gen/stringutils"
-	"github.com/twitchtv/twirp/internal/gen/typemap"
+	"github.com/twitchtv/twirp/exp/gen"
+	"github.com/twitchtv/twirp/exp/gen/stringutils"
+	"github.com/twitchtv/twirp/exp/gen/typemap"
 )
 
 type twirp struct {
